% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R, R/procrustes.R
\name{plot.procrustean}
\alias{plot.procrustean}
\alias{ca_procrustes}
\alias{ca_procrustes.matrix}
\alias{ca_procrustes.incidence_matrix}
\title{Correspondence Analysis with Procrustes Fitting}
\usage{
\method{plot}{procrustean}(result)

ca_procrustes(obj)

\method{ca_procrustes}{matrix}(obj)

\method{ca_procrustes}{incidence_matrix}(obj)
}
\arguments{
\item{obj}{An incidence matrix of size n x k.}
}
\value{
A \code{list} object of class \code{strand} containing the following:
\itemize{
\item \code{ref} The Procrustes-fit coordinates of the scores of the reference seriation.
\item \code{x} The Procrustes-fit coordinates of the row scores of the data.
\item \code{y} The Procrustes-fit coordinates of the column scores of the data.
}
}
\description{
Fit scores of correspondence analysis on an incidence matrix to those produced by reference matrix which contain an ideal seriation using a Procrustes method (on the reference matrix, see \code{\link[lakhesis]{im_ref}}). Rotation is determined by minimizing Euclidean distance from each row score to the nearest reference row score. Correspondence analysis is performed using the \code{\link[ca]{ca}} package \insertCite{nenadic_correspondence_2007}{lakhesis}.
}
\examples{
data("quattrofontanili")
s <- ca_procrustes(quattrofontanili)
# print(s)

}
\references{
\insertAllCited{}
}
