% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion.R
\name{conc_c}
\alias{conc_c}
\alias{conc_c.matrix}
\alias{conc_c.incidence_matrix}
\title{Kendall-Doran (Column) Concentration}
\usage{
conc_c(obj)

\method{conc_c}{matrix}(obj)

\method{conc_c}{incidence_matrix}(obj)
}
\arguments{
\item{obj}{A seriated binary matrix.}
}
\value{
The measure of concentration.
}
\description{
The Kendall-Doran measure of concentration \insertCite{kendall_statistical_1963,doran_computer_1971}{lakhesis}. In a seriated matrix, this function computes the total number cells between the first and last non-zero value, column by column.
}
\examples{
data("quattrofontanili")
conc_c(quattrofontanili)

}
\references{
\insertAllCited{}
}
