% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strand.R
\name{conc_kappa}
\alias{conc_kappa}
\alias{conc_kappa.matrix}
\alias{conc_kappa.incidence_matrix}
\title{Kappa Concentration}
\usage{
conc_kappa(obj)

\method{conc_kappa}{matrix}(obj)

\method{conc_kappa}{incidence_matrix}(obj)
}
\arguments{
\item{obj}{A seriated binary matrix.}
}
\value{
The \eqn{\kappa} coefficient of concentration.
}
\description{
The concentration coefficient \eqn{\kappa}, which extends the Kendall-Doran measure of concentration to include rows and then weights the total measure by the total sum of values in the matrix. See \code{\link[lakhesis]{concentration.col}}.
}
\examples{
data("quattrofontanili")
conc_kappa(quattrofontanili)

}
