% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strand.R
\name{concentration.col}
\alias{concentration.col}
\title{Kendall-Doran Concentration}
\usage{
concentration.col(obj)
}
\arguments{
\item{obj}{A seriated binary matrix.}
}
\value{
The measure of concentration.
}
\description{
The Kendall-Doran measure of concentration \insertCite{kendall_statistical_1963,doran_computer_1971}{lakhesis}. In a seriated matrix, this function computes the total number cells between the first and last non-zero value, column by column.
}
\examples{
# example using Hodson's data from Münsingen-Rain
library(seriation)
data(Munsingen)
munsingen <- Munsingen

# row and column names should be unique
rownames(munsingen) <- paste("Context", rownames(munsingen))
colnames(munsingen) <- paste("Find", colnames(munsingen))

concentration.col(munsingen)

}
\references{
\insertAllCited{}
}
