% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion.R
\name{cor_sp}
\alias{cor_sp}
\alias{cor_sp.matrix}
\alias{cor_sp.incidence_matrix}
\title{Criterion of Optimality: Spearman Correlation}
\usage{
cor_sp(obj)

\method{cor_sp}{matrix}(obj)

\method{cor_sp}{incidence_matrix}(obj)
}
\arguments{
\item{obj}{A seriated binary matrix.}
}
\value{
Spearman's rank correlation coefficient.
}
\description{
Treating each incidence of 1 in an element \eqn{(i,j)} of a seriated matrix as an \eqn{(x,y)} point, computes Spearman's rank correlation coefficient \insertCite{spearman_proof_1904}{lakhesis}.
}
\examples{
data("quattrofontanili")
cor_sp(quattrofontanili)

}
\references{
\insertAllCited{}
}
