% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion.R
\name{cor_sq}
\alias{cor_sq}
\alias{cor_sq.matrix}
\alias{cor_sq.incidence_matrix}
\title{Optimality Criterion: Squared Correlation}
\usage{
cor_sq(obj)

\method{cor_sq}{matrix}(obj)

\method{cor_sq}{incidence_matrix}(obj)
}
\arguments{
\item{obj}{A seriated binary matrix.}
}
\value{
Spearman's rank correlation coefficient.
}
\description{
Treating each incidence of 1 in an element \eqn{(i,j)} of a seriated matrix as an \eqn{(x,y)} point, computes the squared correlation coefficient \insertCite{@see @mccormick_identification_1969, 147-148}{lakhesis}.
}
\examples{
data("quattrofontanili")
cor_sq(quattrofontanili)

}
\references{
\insertAllCited{}
}
