% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strand.R
\name{element.eval}
\alias{element.eval}
\title{Evaluating Element Fit}
\usage{
element.eval(obj)
}
\arguments{
\item{obj}{A seriated binary matrix.}
}
\value{
A list of the \eqn{p} values for each row and column
}
\description{
Performs a deviance test using a quadratic-logistic regression on the rows and columns of a seriated incidence matrix, returning the \eqn{p} value for each row or column. Rows or columns which exhibit perfect separation in 0/1 values are assigned an NA value.
}
\examples{
# using the Münsingen data from the Münsingen-Rain
library(seriation)
munsingen <- data(Munsingen)

# row and column names should be unique
rownames(munsingen) <- paste("Context", rownames(munsingen))
colnames(munsingen) <- paste("Find", colnames(munsingen))

}
