% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\name{im.csv.read}
\alias{im.csv.read}
\title{Read csv File to Incidence Matrix}
\usage{
im.csv.read(
  filename,
  header = FALSE,
  characterencoding = "iso-8859-1",
  remove.hapax = TRUE
)
}
\arguments{
\item{filename}{The filename to uploaded (must be in CSV format).}

\item{header}{If the CSV file contains a header. Default is \code{FALSE}.}

\item{characterencoding}{File encoding as used by \code{\link[readr::locale]{locale}}. Default is \code{"iso-8859-1"} to handle special characters.}

\item{remove.hapax}{Remove any row or column which has a sum of 1 (i.e., is only attested once), since they do not directly contribute to the result of the seriation. Default is \code{TRUE}.}
}
\value{
An incidence matrix of binary values (0 = column element is lacking for row element; 1 = column element is present for row element).
}
\description{
Wrapper around the \code{\link[readr]{read_csv}} function from the \code{\link[readr]{readr}} package \insertCite{wickham_readr_2024}{lakhesis}. Read a CSV file in which the first column represents row elements and the second column represents column elements, and convert it into an incidence matrix.
}
\references{
\insertAllCited{}
}
