% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\name{im.merge}
\alias{im.merge}
\title{Merge Two Incidence Matrices}
\usage{
im.merge(obj1, obj2)
}
\arguments{
\item{obj1, obj2}{Two incidence matrices of any size.}
}
\value{
A single incidence matrix.
}
\description{
From two incidience matrices, create a single incidence matrix. Matrices may contain same row or column elements.
}
