% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\name{im_read_csv}
\alias{im_read_csv}
\title{Read csv File to Incidence Matrix}
\usage{
im_read_csv(
  filename,
  header = FALSE,
  characterencoding = "iso-8859-1",
  remove.hapax = FALSE
)
}
\arguments{
\item{filename}{The filename to uploaded (must be in \code{.csv} format).}

\item{header}{If the \code{.csv} file contains a header. Default is \code{FALSE}.}

\item{characterencoding}{File encoding as used by \code{\link[readr]{locale}}, which depends on the \code{csv} file. Default is \code{"iso-8859-1"} but \code{"utf-8"} is also recommended.}

\item{remove.hapax}{Remove any row or column which has a sum of 1 (i.e., is only attested once), since they do not directly contribute to the result of the seriation. Default is \code{FALSE}.}
}
\value{
A matrix of binary values (0 = row/column occurrence is absence; 1 = row/column occurrence is present).
}
\description{
Wrapper around the \code{\link[readr]{read_csv}} function from the \code{readr} package \insertCite{wickham_readr_2024}{lakhesis}. Read a \code{.csv} file in which the first column represents row elements and the second column represents column elements, and convert it into an incidence matrix.
}
\references{
\insertAllCited{}
}
