% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strand.R
\name{kappa.coef}
\alias{kappa.coef}
\title{Kappa Concentration}
\usage{
\method{kappa}{coef}(obj)
}
\arguments{
\item{obj}{A seriated binary matrix.}
}
\value{
The \eqn{\kappa} coefficient of concentration.
}
\description{
The concentration coefficient \eqn{\kappa}, which extends the Kendall-Doran measure of concentration to include rows, and then weights the total measure by the total sum of the matrix. See \code{\link[lakhesis]{concentration.col}}.
}
\examples{
# using the Münsingen data from the Münsingen-Rain
library(seriation)
munsingen <- data(Munsingen)

# row and column names should be unique
rownames(munsingen) <- paste("Context", rownames(munsingen))
colnames(munsingen) <- paste("Find", colnames(munsingen))

kappa.coef(munsingen)

}
