% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakhesize.R
\name{lakhesize}
\alias{lakhesize}
\title{Lakhesize}
\usage{
lakhesize(strands, obj)
}
\arguments{
\item{strands}{A list of \code{strands}, which are data frames returned by \code{\link[lakhesis]{ca.procrustes.curve}}.}

\item{obj}{The intial incidence matrix.}
}
\value{
A list of the following:
\itemize{
\item \code{RowConsensus} Data frame of the consensus seriation of the row elements in the order of their projection on the first principal axis.
\item \code{ColConsensus} Data of the consensus seriation of the column elements in the order of their project onto the first principal axis.
\item \code{RowPCA} The results of \code{prcomp} performed on the regressed row elements of strands.
\item \code{ColPCA} The results of \code{prcomp} performed on the regressed column elements of strands.
\item \code{Coef}  Coefficients for each strand:
\itemize{
\item Agreement: the measure of how well each strand accords with the consensus seriation, \eqn{\rho_r}\eqn{\rho)c},  the product of the square of Spearman's rank correlation coefficient, \eqn{\rho}, of the row and column strand with those of consensus.
\item Concentration: the concentration coefficient \eqn{\kappa} (see \code{\link[lakhesis]{kappa.coef}}).
}
}
}
\description{
This function returns the row and column consensus serations for a data frame of strands, in the form of a \code{list} of their rankings,
the results of their regression, and coefficients of agreement and concentration for each strand. See documentation
on \code{\link[lakhesis]{kappa.coef}} and  \code{\link[lakhesis]{spearman.sq}}.
}
