% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{LC}
\alias{LC}
\title{Lakhesis Calculator}
\usage{
LC()
}
\value{
Opens the Lakhesis Calculator. CSV files are imported (see \link[lakhesis]{im.csv.read} for formatting).
Results are download as an \code{/rds} file containing a list object \code{results} of  the following:
\itemize{
\item \code{results} The consensus seriation achieved, including coefficients of agreement and concentration
(\code{\link[lakhesis]{lakhesize}}).
\item \code{strands} The strands selected by the investigator.
\item \code{im.seriated} The incidence matrix of the consensus seriation.
}
}
\description{
Launch Lakhesis Calculator, a graphical interface to explore binary matrices via correspondence analysis,
select potentially well-seriated sequences, and perform consensus seriation.
Interface is made with \code{ggplot2}, \code{shiny}, \code{shinydashboard}, \code{bslib}
\insertCite{wickham_ggplot2:_2016,chang_shiny_2024,chang_shinydashboard_2021,sievert_bslib_2024}{lakhesis}.
}
\references{
\insertAllCited{}
}
