% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qf_strands}
\alias{qf_strands}
\title{Quattro Fontanili - Strands}
\format{
A \code{strands} object containing strands output by \code{\link[lakhesis]{ca_procrustes}}.
}
\usage{
data("qf_strands")
}
\description{
Three seriated strands selected from \code{\link[lakhesis]{quattrofontanili}} data, identified by the package author as an example for the documentation of functions.
}
\examples{
data("qf_strands")
print(qf_strands)

}
\keyword{datasets}
