% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qfLakhesis}
\alias{qfLakhesis}
\title{Quattro Fontanili - Lakhesis Results}
\format{
A list containing the following objects output by \code{\link[lakhesis]{lakhesize}}.
}
}
\usage{
data("qfLakhesis")
}
\description{
The results of \code{\link[lakhesis]{lakhesize}} applied to \code{\link[lakhesis]{quattrofontanili}} data (exported and reimported into the Calculator). Seven strands were selected by the package author as an example for the documentation of functions. Consult the \code{\link[lakhesis]{lakhesize}} function for the structure of the object.
}
\examples{
data("qfLakhesis")
print(qfLakhesis)

}
\keyword{datasets}
