% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qfStrands}
\alias{qfStrands}
\title{Quattro Fontanili - Strands}
\format{
A list containing data frames output by \code{\link[lakhesis]{ca.procrustes.curve}}.
}
\usage{
data("qfStrands")
}
\description{
Seven seriated strands selected from \code{\link[lakhesis]{quattrofontanili}} data, identified by the package author as an example for the documentation of functions.
}
\examples{
data("qfStrands")
print(qfStrands)

}
\keyword{datasets}
