% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quattrofontanili}
\alias{quattrofontanili}
\title{Quattro Fontanili}
\format{
A seriated incidence matrix of 81 rows (tombs) and 82 columns (types).
\describe{
Data entered from \insertCite{close-brooks_veii_1979;textual}{lakhesis}, an English translation of the authors' original publication in \emph{Notizie degli Scavi} (1963). Descriptions of types may be found in that paper.
}
}
\usage{
data("quattrofontanili")
}
\description{
The seriation of tombs from necropoleis at Veii, primarily Quattro Fontanili, but also Valle la Fata, Vaccareccia, and Picazzano, in southern Etruria, established by \insertCite{close-brooks_veii_1979;textual}{lakhesis}.
}
\examples{
data("quattrofontanili")
print(quattrofontanili)

}
\references{
\insertAllCited{}
}
\keyword{datasets}
