% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strand.R
\name{strand_create}
\alias{strand_create}
\alias{strand_create.matrix}
\alias{strand_create.incidence_matrix}
\title{Create Strand Object from Seriated Incidence Matrix}
\usage{
strand_create(obj, method = NULL)

\method{strand_create}{matrix}(obj, method = NULL)

\method{strand_create}{incidence_matrix}(obj, method = NULL)
}
\arguments{
\item{obj}{A \code{list} of strands.}

\item{method}{The method used to create the strand (optional).}
}
\value{
A \code{list} of class \code{strands}.
}
\description{
Given a seriated incidence matrix with unique row and column names, create a \code{strand} object.
}
