% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strand.R
\name{strands_create}
\alias{strands_create}
\alias{strands_create.list}
\title{Create List of Strands}
\usage{
strands_create(strands)

\method{strands_create}{list}(strands)
}
\arguments{
\item{strands}{A \code{list} of strands.}
}
\value{
A \code{list} of class \code{strands}.
}
\description{
Given one or more individual strand objects, create a single \code{list} of class \code{strands}.
}
