% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strand.R
\name{conc_col}
\alias{conc_col}
\title{Kendall-Doran Concentration}
\usage{
conc_col(obj)
}
\arguments{
\item{obj}{A seriated binary matrix.}
}
\value{
The measure of concentration.
}
\description{
The Kendall-Doran measure of concentration \insertCite{kendall_statistical_1963,doran_computer_1971}{lakhesis}. In a seriated matrix, this function computes the total number cells between the first and last non-zero value, column by column.
}
\examples{
data("quattrofontanili")
conc_col(quattrofontanili)

}
\references{
\insertAllCited{}
}
