% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\name{im.merge}
\alias{im.merge}
\title{Merge Two Incidence Matrices}
\usage{
im.merge(obj1, obj2)
}
\arguments{
\item{obj1, obj2}{Two incidence matrices, which may contain overlapping row or column values.}
}
\value{
A single incidence matrix.
}
\description{
Create an ideal reference matrix of well-seriated values of the same size as the input matrix.
}
