% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakhesize.R, R/print.R
\name{lakhesize}
\alias{lakhesize}
\alias{lakhesize.strands}
\alias{lakhesize.default}
\alias{plot.lakhesis}
\title{Lakhesize}
\usage{
lakhesize(strands, ...)

\method{lakhesize}{strands}(strands, pbar = TRUE)

\method{lakhesize}{default}(strands, pbar = TRUE)

\method{plot}{lakhesis}(result, display = "im_seriated")
}
\arguments{
\item{strands}{A \code{list} of the \code{strands} class (see \code{\link[lakhesis]{add_strand}}).}

\item{pbar}{Displaying a progress bar. Default is \code{TRUE}.}
}
\value{
A \code{list} of class \code{lakhesis} containing the following:
\itemize{
\item \code{row} A seriated vector of row elements.
\item \code{col} A seriated vector of column elements
\item \code{coef}  A \code{data frame} containing the following columns:
\itemize{
\item \code{Strand} The number of the strand.
\item \code{Agreement} The measure of agreement, i.e., how well each strand accords with the consensus seriation. Using the square of Spearman's rank correlation coefficient, \eqn{\rho^2}, between each strand and the consensus ranking, agreement is computed as the product of \eqn{\rho^2} for their row and column rankings, \eqn{\rho_r^2}\eqn{\rho_c^2}.
\item \code{Concentration} the concentration coefficient \eqn{\kappa}, which provides a measure of the optimality of each strand (see \code{\link[lakhesis]{kappa.coef}}).
}
\item \code{im_seriated} The seriated incidence matrix, of class \code{incidence_matrix}.
}
}
\description{
This function returns the row and column consensus seriation for a \code{list} object of the \code{strands} class, containing their rankings and coefficients of association and concentration.
}
\details{
Consensus seriation is achieved by iterative simple linear regression to handle \code{NA} vales in each strand. To initialize, a regression is performed pairwise, with every strand as the dependent \eqn{y} variate and every other strand as the independent \eqn{x} variate. The independent variate's rankings are then regressed onto \eqn{f(x) = \hat{\beta}_1 x + \hat{\beta}_0}. If \eqn{y \neq f(x)}, the mean of \eqn{y} and \eqn{f(x)} is used. Then, the values of dependent variate and those of regressed independent are re-ranked, which serves as the dependent variate on the next iteration. The process is repeated, regressing each strand which yields the lowest concentration measure.
}
\examples{
data("qfStrands")
x <- lakhesize(qfStrands, pbar = FALSE)
# summary(x) 

}
