% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakhesize.R, R/print.R
\name{lakhesize}
\alias{lakhesize}
\alias{lakhesize.strands}
\alias{lakhesize.default}
\alias{plot.lakhesis}
\title{Lakhesize}
\usage{
lakhesize(strands, ...)

\method{lakhesize}{strands}(strands, iter = 100, sim = 1, pbar = TRUE)

\method{lakhesize}{default}(strands, iter = 100, sim = 1, pbar = TRUE)

\method{plot}{lakhesis}(result, display = "im_seriated")
}
\arguments{
\item{strands}{A \code{list} of the \code{strands} class (see \code{\link[lakhesis]{add_strand}}).}

\item{iter}{The number of iterations to use in lakhesizing the \code{strands}. The default is set at 100.}

\item{sim}{The number of simulation runs of iterations. The default is set at 1. Increasing the number of simulations will better ensure an optimal ordering.}

\item{pbar}{Displaying a progress bar. Default is \code{TRUE}.}

\item{obj}{The intial incidence matrix.}
}
\value{
A \code{list} of class \code{lakhesis} containing the following:
\itemize{
\item \code{row} A seriated vector of row elements.
\item \code{col} A seriated vector of column elements
\item \code{rowPCA} The results of \code{\link[stats]{prcomp}} performed on the row elements of \code{strands}.
\item \code{colPCA} The results of \code{\link[stats]{prcomp}} performed on the column elements of \code{strands}.
\item \code{coef}  A \code{data frame} containing the following columns:
\itemize{
\item \code{Strand} The number of the strand.
\item \code{Agreement} The measure of agreement, i.e., how well each strand accords with the consensus seriation. Using the square of Spearman's rank correlation coefficient, \eqn{\rho^2}, between each strand and the consensus ranking, agreement is computed as the product of \eqn{\rho^2} for their row and column rankings, \eqn{\rho_r^2}\eqn{\rho_c^2}.
\item \code{Concentration} the concentration coefficient \eqn{\kappa}, which provides a measure of the optimality of each strand (see \code{\link[lakhesis]{kappa.coef}}).
}
\item \code{im_seriated} The seriated incidence matrix, of class \code{incidence_matrix}.
}
}
\description{
This function returns the row and column consensus seriation for a \code{list} object of the \code{strands} class, containing their rankings, the results of their PCA, and coefficients of association and concentration.
}
\details{
Consensus seriation is achieved by iterative linear regression using random simulation. On one iteration, strands are chosen at random, omitting incomplete or missing pairs, using PCA to determine the best-fitting line for their rankings. Both strands' rankings are then regressed onto that line to determine missing values, and then re-ranked, repeating until all strands have been regressed. PCA of the simulated rankings is then used to determine the final sequence of the row and column elements.

A default setting of 100 iterations (\code{iter = 100, sim = 1}) is used in the Lakhesis calculator (\code{\link[lakhesis]{LC}}). Increasing simulations is recommended for post-exploratory analysis, to determine if a more optimal seriations can be found than the estimate yielded from the initial "exploratory" result. The number of iterations (\code{iter}) and simulation runs of iterations (\code{sim}) can be increased. The simultation run which yields the lowest (most optimal) measure of concentration \eqn{\kappa} (see \code{\link[lakhesis]{conc_kappa}}) chosen as the output.
}
\examples{
data("qfStrands")
x <- lakhesize(qfStrands, pbar = FALSE)
# summary(x) 

}
